legacy_versions/LDS/requirements.txt
# Core Django Framework
Django>=4.2

# Django REST Framework for API development
djangorestframework==3.15.2

# CORS headers for cross-origin requests
django-cors-headers==4.6.0

# Code Analysis and Parsing
javalang==0.13.0

# Diagram Generation
graphviz==0.20.3

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0

# Code Quality
pylint==3.3.2
black==24.10.0
flake8==7.1.1

# Type Checking
mypy==1.13.0

# Utilities
python-dotenv==1.0.1

# AI Documentation Generation
google-generativeai
Jinja2
markdown
xhtml2pdf
